# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.FriendshipRequestResponse:
    properties:
      blockedAt:
        type: string
      createdAt:
        type: string
      id:
        type: string
      isBlocked:
        type: boolean
      isReceivedByUser:
        type: boolean
      isSentByUser:
        type: boolean
      recipientAvatar:
        type: string
      recipientName:
        type: string
      senderAvatar:
        type: string
      senderName:
        type: string
      updatedAt:
        type: string
    type: object
  dto.FriendshipResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      profileAvatar:
        type: string
      profileId:
        type: string
      profileName:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.ItemParticipantRequest:
    properties:
      profileId:
        type: string
      weight:
        type: integer
    required:
    - profileId
    type: object
...
  dto.NewExpenseItemRequest:
    properties:
      amount:
        type: number
      name:
        minLength: 3
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - amount
    - name
    - quantity
    type: object
...
  dto.SyncItemParticipantsRequest:
    properties:
      participants:
        items:
          $ref: '#/definitions/dto.ItemParticipantRequest'
        type: array
    type: object
...
  dto.UpdateExpenseItemRequest:
    properties:
      amount:
        type: number
      name:
        minLength: 3
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - amount
    - name
    - quantity
    type: object
...
  response.JSONResponse-array_dto_FriendshipRequestResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.FriendshipRequestResponse'
        type: array
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.JSONResponse-dto_FriendshipResponse:
    properties:
      data:
        $ref: '#/definitions/dto.FriendshipResponse'
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.Pagination:
    properties:
      currentPage:
        type: integer
      hasNextPage:
        type: boolean
      hasPrevPage:
        type: boolean
      totalData:
        type: integer
      totalPages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Cashback backend API
  title: Cashback API
  version: "1.0"
paths:
...
  /friend-requests/{type}:
    get:
      parameters:
      - description: 'Request type: sent or received'
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-array_dto_FriendshipRequestResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all friend requests (sent or received)
      tags:
      - friend-requests
  /friend-requests/received/{requestId}:
    delete:
      parameters:
      - description: Friend request ID
        in: path
        name: requestId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Ignore a received friend request
      tags:
      - friend-requests
    patch:
      parameters:
      - description: Friend request ID
        in: path
        name: requestId
        required: true
        type: string
      - description: 'Command: block or unblock'
        in: query
        name: command
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Block or unblock a friend request sender
      tags:
      - friend-requests
    post:
      parameters:
      - description: Friend request ID
        in: path
        name: requestId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_FriendshipResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Accept a received friend request
      tags:
      - friend-requests
  /friend-requests/sent/{requestId}:
    delete:
      parameters:
      - description: Friend request ID
        in: path
        name: requestId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Cancel a sent friend request
      tags:
      - friend-requests
...
  /group-expenses/{groupExpenseId}/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      - description: New expense item payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.NewExpenseItemRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add an item to a group expense
      tags:
      - expense-items
  /group-expenses/{groupExpenseId}/items/{expenseItemId}:
    delete:
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      - description: Expense item ID
        in: path
        name: expenseItemId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove an expense item
      tags:
      - expense-items
    put:
      consumes:
      - application/json
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      - description: Expense item ID
        in: path
        name: expenseItemId
        required: true
        type: string
      - description: Update expense item payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateExpenseItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update an expense item
      tags:
      - expense-items
  /group-expenses/{groupExpenseId}/items/{expenseItemId}/participants:
    put:
      consumes:
      - application/json
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      - description: Expense item ID
        in: path
        name: expenseItemId
        required: true
        type: string
      - description: Participants payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SyncItemParticipantsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Sync participants of an expense item
      tags:
      - expense-items
...
  /profiles/{profileId}/friend-requests:
    post:
      parameters:
      - description: Target profile ID
        in: path
        name: profileId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Send a friend request
      tags:
      - friend-requests
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"